backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
anthropic>=0.68.2
google-genai>=1.0.0
httpx>=0.27.0
beautifulsoup4>=4.12.3
lxml>=5.0.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
pydantic>=2.7.0
python-dotenv>=1.0.1
aiofiles>=23.2.1
